+++ grow/grow-delay/grow-delay-redisson-boot-starter/src/main/java/com/liguanqiao/grow/delay/redisson/spring/boot/autoconfigure/GrowDelayAutoConfiguration.java
@@ -5,7 +5,7 @@ import com.liguanqiao.grow.delay.handler.DelayTaskHandler;
 import com.liguanqiao.grow.delay.redisson.DelayTaskOpsRedissonImpl;
 import com.liguanqiao.grow.delay.redisson.config.RedissonDelayQueueRegistry;
 import com.liguanqiao.grow.log.context.TracerContext;
-import com.liguanqiao.grow.log.context.TracerContextDefaultImpl;
+import com.liguanqiao.grow.log.util.TracerContextUtil;
 import lombok.extern.slf4j.Slf4j;
 import org.redisson.api.RedissonClient;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,9 +14,6 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 import java.util.List;
-import java.util.Optional;
-
-;
 
 /**
  * @author liguanqiao
@@ -28,12 +25,12 @@ public class GrowDelayAutoConfiguration {
 
     @Bean
     @ConditionalOnMissingBean
-    public DelayTaskOps delayTaskOps(RedissonClient redissonClient, @Autowired(required = false) TracerContext tracerContext, List<DelayTaskHandler<?>> handlers) {
+    public DelayTaskOps delayTaskOps(RedissonClient redissonClient, List<DelayTaskHandler<?>> handlers, @Autowired(required = false) TracerContext tracerContext) {
         log.info(">>>>>>>>>>> Grow DelayTask Redisson Config Init.");
-        TracerContext tracerContext1 = Optional.ofNullable(tracerContext).orElseGet(TracerContextDefaultImpl::new);
+        TracerContext tc = TracerContextUtil.getOrDefault(tracerContext);
         RedissonDelayQueueRegistry registry = new RedissonDelayQueueRegistry(redissonClient, handlers);
-        DelayTaskOpsRedissonImpl delayTaskOps = new DelayTaskOpsRedissonImpl(registry, tracerContext1);
-        registry.registerAllListener(delayTaskOps, tracerContext1);
+        DelayTaskOpsRedissonImpl delayTaskOps = new DelayTaskOpsRedissonImpl(registry, tc);
+        registry.registerAllListener(delayTaskOps, tc);
         return delayTaskOps;
     }