+++ grow/grow-mq/grow-mq-rabbit-boot-starter/src/main/java/com/liguanqiao/grow/mq/spring/boot/autoconfigure/GrowMqAutoConfiguration.java
@@ -1,10 +1,10 @@
 package com.liguanqiao.grow.mq.spring.boot.autoconfigure;
 
 import com.liguanqiao.grow.log.context.TracerContext;
-import com.liguanqiao.grow.log.context.TracerContextDefaultImpl;
+import com.liguanqiao.grow.log.util.TracerContextUtil;
 import com.liguanqiao.grow.mq.MqSender;
+import com.liguanqiao.grow.mq.rabbit.GrowRabbitMessageConverter;
 import com.liguanqiao.grow.mq.rabbit.MqSenderRabbitImpl;
-import com.liguanqiao.grow.mq.rabbit.RabbitMessageConverter;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.amqp.core.AmqpTemplate;
 import org.springframework.amqp.support.converter.MessageConverter;
@@ -13,8 +13,6 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.annotation.Order;
-
-import java.util.Optional;
 
 /**
  * @author liguanqiao
@@ -29,12 +27,12 @@ public class GrowMqAutoConfiguration {
     @Bean
     public MqSender mqSender(AmqpTemplate template, @Autowired(required = false) TracerContext tracerContext) {
         log.info(">>>>>>>>>>> Grow MqSender Rabbit Config Init.");
-        return new MqSenderRabbitImpl(template, Optional.ofNullable(tracerContext).orElseGet(TracerContextDefaultImpl::new));
+        return new MqSenderRabbitImpl(template, TracerContextUtil.getOrDefault(tracerContext));
     }
 
     @Bean
     @ConditionalOnMissingBean
     public MessageConverter rabbitMessageConverter(@Autowired(required = false) TracerContext tracerContext) {
-        return new RabbitMessageConverter(Optional.ofNullable(tracerContext).orElseGet(TracerContextDefaultImpl::new));
+        return new GrowRabbitMessageConverter(TracerContextUtil.getOrDefault(tracerContext));
     }
 }